DBZ-252 Exclude reusable logic from LegacyDdlParser
